requirements.txt
win10toast
pypiwin32
setuptools
infi.systray
requests